Separated RegolithSpawner into its own files  - Changed node name to regolith_node
